+++ src/filesystem/network/File.h
@@ -33,7 +33,7 @@ class NetworkFile : public CommonFile
 public:
     NetworkFile( const std::string& mrl );
     virtual unsigned int lastModificationDate() const override;
-    virtual unsigned int size() const override;
+    virtual int64_t size() const override;
     virtual bool isNetwork() const override;
 };
 }
