+++ src/filesystem/unix/File.cpp
@@ -48,7 +48,7 @@ unsigned int File::lastModificationDate() const
     return m_lastModificationDate;
 }
 
-unsigned int File::size() const
+int64_t File::size() const
 {
     return m_size;
 }
